RAG_FINANCIAL_PDF/requirements.txt
# Core RAG and Vector Database
qdrant-client>=1.7.0
sentence-transformers>=2.3.1

# LLM and API
requests>=2.31.0
httpx>=0.25.0

# Web Interface
streamlit>=1.28.0

# Data Processing
pypdf>=3.17.4
python-dotenv>=1.0.0

# Utilities
pydantic>=2.5.0
loguru>=0.7.0
